[FIX] web_kanban: double click double action
When multiple clicking fast (and with the server not too fast), we could
end up with an error (eg. "Cannot read property 'controller' of null").

Something has been done previously to improve this when switching
between views in: e2781cd6. This commit do something similar to limit
multiple click on kanban records (doing it as in 9f5bedf0a).

opw-705329
closes #16698