[FIX] point_of_sale: don't reserve pickings with tracked products
Introduced by 5ad4a394b953e6985435bf93a4ac215047c891c3 which was
intended for 9.0. It does not work well in 10.0 because here users
have the possibility to specify their own tracking numbers directly in
the POS interface. This causes issues when eg. the user specifies the
same lot number as is already reserved by the system.

opw-742657