[FIX] base: ir_cron active_test in context
When the list view of 'ir_cron' is opened, the key 'active_test' is set
to False in the context to see all cron, even the archived ones.

The problem by passing 'active_test' in context is that it's propagated
to all actions launched from this view, for example whan you run
manually a cron, this key is passed to the cron action, which may leads
to call some actions on archived records which is not intended.

To fix this issue, we removed the key 'active_test' from context and
instead create a filter that show all crons set by default.