[FIX] pos_mercury: take into account rescue sessions
If there are open rescue sessions there will be multiple opened
pos.session records with the same user.

A better solution would be to make the POS client send back the
pos.session id it belongs to and use that instead, but this would be
unsuitable for a stable release. It will instead be done in master.

opw-743318